:truck: Move .spec.js from test to src
