+++ src/conceptNetworkState.spec.js
@@ -1,6 +1,6 @@
 import from from 'from';
 import ezs from 'ezs';
-import statements from '../src';
+import statements from '.';
 import { ConceptNetworkState } from 'concept-network';
 
 ezs.use(statements);
